Update model_storage_usage.go
Increase StorageUsage.bytes from int32 to int64
See #5

+++ model_storage_usage.go
@@ -18,7 +18,7 @@ import (
 
 // StorageUsage struct for StorageUsage
 type StorageUsage struct {
-	Bytes     *int32     `json:"bytes"`
+	Bytes     *int64     `json:"bytes"`
 	UpdatedAt *IonosTime `json:"updatedAt,omitempty"`
 }
